feat(tracking): move some code out of service
