+++ packages/backend/src/helpers/tracking.js
@@ -1,0 +1,23 @@
+const actions = {
+  creation: "CREATION",
+  deactivation: "DEACTIVATION",
+  deletion: "DELETION",
+  modification: "MODIFICATION",
+};
+
+module.exports.actions = actions;
+
+const entities = {
+  eig: "EIG",
+  userBack: "USER_BACK",
+  userFront: "USER_FRONT",
+};
+
+module.exports.entities = entities;
+
+const userTypes = {
+  back: "BACK",
+  front: "FRONT",
+};
+
+module.exports.userTypes = userTypes;
