+++ packages/backend/src/routes/user.js
@@ -5,7 +5,7 @@ const router = express.Router();
 const checkJWT = require("../middlewares/checkJWT");
 const userController = require("../controllers/user");
 const trackFoUser = require("../middlewares/trackFoUser");
-const { actions, userTypes } = require("../services/Tracking");
+const { actions, userTypes } = require("../helpers/tracking");
 
 // Gère une connexion via mot de passe.
 router.get("/me", checkJWT, userController.getMe);
